Overlay moved to separate component
